Added mailers and google sign in

+++ Controllers/api/v1/posts_api.js
@@ -30,27 +30,30 @@ module.exports.destroy = async function (req, res) {
     try {
         let post = await Post.findById(req.params.id);
 
-        // if (post.user == req.user.id) {
-        await post.remove();
-        await Comment.deleteMany({ post: req.params.id });
+        if (post.user == req.user.id) {
+            await post.remove();
+            await Comment.deleteMany({ post: req.params.id });
 
-        return res.json(200, {
-            mesage: 'Post and associated comments deleted successfully'
-        });
+            return res.json(200, {
+                mesage: 'Post and associated comments deleted successfully'
+            });
 
-        // if (req.xhr) {
-        //     return res.status(200).json({
-        //         data: {
-        //             post_id: req.params.id
-        //         },
-        //         message: "Post deleted"
-        //     });
-        // }
+            // if (req.xhr) {
+            //     return res.status(200).json({
+            //         data: {
+            //             post_id: req.params.id
+            //         },
+            //         message: "Post deleted"
+            //     });
+            // }
 
-        // req.flash('success', 'Deleted the comments :(');
-        // } else {
-        // req.flash('error', 'Comments not deleted unauthorized :(');
-        // }
+            // req.flash('success', 'Deleted the comments :(');
+        } else {
+            // req.flash('error', 'Comments not deleted unauthorized :(');
+            return res.json(401, {
+                message: 'Unauthorized'
+            });
+        }
     } catch (err) {
         console.log(err)
         // req.flash('error', err);
